[PP] Add first implementation of a PlatformProxy module
This is the first instance of a PlatformProxy, which is the Barbeque core
module in charge to interface the RTRM with the specific target platform.
The initial implementation, still to be completed, defines the basic services
exposed to other modules, which allows to:
1. query for the resources available on the target platform
2. assign resources to an application
3. reclaim resources from an application

The PlatformProxy module will provide also a proper platform monitoring service
which could be used to trigger a RTRM action once resources availability
should change at run-time.

The PlatformPoxy module is meant to be extended by a platform specific instance,
to be implemented as a derived class extending from this basic module which
instead is used to provide the platform independent code to access a generic
target platform.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>