[PP_LINUX] Add first implementation of a Linux PlatformProxy
This is the first instance of a PlatformProxy for a generic Linux host which
exploits the Control Groups framework to constraints CPUs and MEMs usage
for the Barbeque controlled applications.

This initial version is still missing the platform resources discovery support,
while an initial working support is provided to MAP and RECLAIM resources
to scheduled applications. The current implementation:
0. setup a "SILOS" control group, with reduced resources, to host applications
READY but without assigned resource;
1. build a control group for each new registered application;
2. setup the application-specific control group for a specified (active) application,
which is based on the scheduling decision;
3. move applications from the SILOS to the application specific CGroup once
required by a user module (e.g. the SynchronizationManager)

Signed-off-by: Patrick Bellasi <derkling@gmail.com>