[YM] Link the PlatformProxy module
This patch exploit the PlatformProxy to enforce resource assignment on active
applications which have been scheduled to run on the target platform.
Platform resources enforcement is required in between the SYNC_CHANGE e the
corresponding DO_CHANGE. Indeed, in this point, applications are expected to
be suspended in a synchronization point. Thus, eventually previously assigned
resources could be reclaimed and new one assigned, right before the application
is allowed to continue.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>