[PP_LINUX] Add support to load CGroup defines resources
When running Barbeque to manage resources of a generic Linux host, we exploit
the Control Group framework not only to assign resources to scheduled applications
but also the collect a description of Barbeque managed resources.

When the BarbequeRTRM start, the PlatformProxy trigger a "resource enumeration"
task which is in charge to identify the set of resources under control of
the run-time manager and register them to the ResourceAccounter.
This resource enumeration activity, when targeting the run-time management
of a generic Linux host, is based on the Control Gropup framework.

Barbeque expects a "bbq_resources" control group being defined with a "cpuset"
controller attached. This control group must define the set of CPUs and memory
nodes which are under control of the run-time resource manager.
CPUs and MEMs could be partitioned into "nodes", e.g. to properly map the
resource hierarchy of a NUMA machine. To support such a partitioning into nodes
we expect to find a set of sub-groups under "bbq_resources", one for each node
into witch resources should be partitioned.
For example, to assign Barbeque the management of 3 clusters of a NUMA machine,
each one with 4 CPUs and 1 memory, it is expected a partitioning like this:

bbq_resources:
  +--- cpus 4-15
  +--- mems 1-3
  +--- node0:
         +--- cpus 1-7
         +--- mems 1
  +--- node1:
         +--- cpus 8-11
         +--- mems 2
  +--- node2:
         +--- cpus 12-15
         +--- mems 3

This patch provide a complete support collect the partitioning of CPUs into
clusters, while the assignement of memory nodes is still missing.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>