[PP_LINUX] Update BBQ resource CGroup
The definition of the CGroup used to configure the BBQ managed resources
has been updated to be also the root for application specific groups.
This allows to enforce a kernel checking on not assigning applications more
resources than the one reserved to BBQ.

Moreover, by properly configuring the top-level CGroup, this allows to
assign resources to applications without requiring BBQ to run with SYS
capabilities.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>