[PP_LINUX] Define the generic LINUX platform
A generic Linux platform will return the "org.linux.cgroup" string identifier.
This is the identifier which is expected on the parsing of recipes file
to load a set of AWM for each application.

This is the identifier of a generic Linux host platform where resources are
controlled by exploiting the kernel Control Group framework.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>