[AEM_CPP] Add onResume callback
This new API is meant to be extended by application code in order to recovery
from a suspension, thus allowing a complete implementation of the
suspend/resume functionality, which is based on these two methods:

.:: onSuspend
Once Barbeque need to stop an application, the onSuspend application defined
method is called, which role is to give the chance the application to
properly release all resources not needed during the suspended state.
A proper implementation requires an application to close all handlers
(e.g. files, sockets) and release HW resources (e.g. close a camera device).
At exit from this call the application is expected to have, as unique
resource still in use, the minimum amount of memory required to save its
status.

.:: onResume
Once Barbeque has the chance to resume a previously suspended application,
the onResume application defined method is called, which goal is to give it
the chance to recover its working state.
A proper implementation requires an application to acquire all the resources
needed by the following onConfigure call.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>