[YM] Bugfix commit on not notified applications
During the last step of the synchronization protocol, i.e. the PostChange,
all application must be committed. Otherwise, they are not removed from the
synchronization queues and the policy could enter an endless loop.

For example, reshuffled AWMs could be not notified to applications.
In this case, the DoSync policy call will return a false but the application
must be anyway removed from the RECONF queue since, at this SyncProto point,
the PlatformProxy has laready provided a correct reshuffling.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>