[Daemon] [PIL-Parser 2/2] Switch to PIL Layout descriptors
The Layout of BBQ managed resources is now independent of the CGroup
configuration file. This patch introduces the new "Barbeque Platform Layout"
file format, which have extension ".bpl" and replaces the previously used
"*.cgroup" files.

An example configuration, suitable for testing on a generic 4core system, is
also provided with a complete description of the configuration syntax

Signed-off-by: Patrick Bellasi <derkling@gmail.com>