[P2012_PP] Initial version
This a first implementation (not complete) of the ST p2012 platform proxy.
Please note that methods for monitoring and platform notification are
currently missing.

To be instanced by the PlatformProxy class, BBQUE_PLATFORM_2012 must
be defined.

The exchange of messages between BarbequeRTRM and P2012 (in the real
exectution mode) is following described:

- BBQ_START: Notify the platform about BBQ starting
- BBQ_UPDATE: There has been a change in resource assignments. In the
  message payload is specified the index of the EXC constraints
  descriptor involved.
- BBQ_STOP: BBQ is stopping. Eventually used for a graceful stop.

- P2012_EVENT: The platform signals an HW event (change of resoruce
  availability, power/thermal issues...)
- P2012_OFFLINE: The platform confirm the reception of the stop message
  sent before. This is used to stop the execution of the monitoring
thread.

Information regarding the resource partitioning defined by the scheduling
policy are written into a shared memory buffer, containing the descriptor
of the managed device (P2012).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>