[Tools_OLPB] Add support for DMMLib tuning knobs
A second file is produced in output which collects all the DMM Library tuning
knobs parameters. The Dynamic Memory Manager (DMM) libray:
https://bitbucket.org/joko/dmmlib
provides support for run-time tuning of dynamic memory allocation policies.
A set of tuning parameters could be used to configure the behavior of the
library and the BarbequeRTRM, actually the RTLib, could hooks into a set of
DMMLib tuning functions in order to setup the library behavior according to
the currently selected AWM.

Tuning parameters could be added to the input Operating Point XML file,
which is produced by a DSE tool (e.g. MOST). All the parameters which name
starts with "dmm." are used to initialize a vector of structs.
The layout of there structs is expected to be defined by the DMM library
provided <dmm/knobs.h> header file; the number and position of struct members
is expecte to match with that found on the input XML file.
is expected to be the

Signed-off-by: Patrick Bellasi <derkling@gmail.com>