[BuildingSystem] Reorganize to be more modular
The layout of this CMake fragment has been reorganized into section in order
to improve readability and be more modular.
Source files and library dependencies are collected into a dedicated variable,
in a dedicated section, based on KConfig configuration options.

A target specific section is added at the end. This target section could be
replicated by adding more files and/or dependencies to produce different
targets.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>