[RTLIB_DMM] Add first implementation
This is the first implementation of the interface layer toward the Dynamic
Memory Management Library (DMMLib) by ICCS:
https://bitbucket.org/joko/dmmlib

This integration is based on a set of symbols which are expected to be
exported by the library:
- dmm_knobs: is an array of knobs parameters for a run-time configuration;
- dmm_knobs_count: the number of run-time configurations
- dmm_rtm_init: initialize DMMLib for run-time management
- dmm_set_knobs: tune DMMLib according to the specified set of knob parameters;

These symbols are "weak", thus if the library should not export them the
integration layer is disabled.

The set of knobs parameters, entries of the dmm_knobs array, are application
specific and could be generated using the bbque-oplb tool.
The Operating-Point List Builder is a parser of XML files produced for example
by the MOST Design Space Exploration tool. An example of such a file is provided
in tools/inputTest_bbqueOpListBuilder.xml.
All the "parameters" and "system_metrics" which names starts with "dmm." are
used to populate the dmm_knobs array.
NOTE: the ordinal position of parsed parameters is expected to match the
layout of the "struct dmm_knobs_t", which is defined by the DMMLib in the
provided <dmm/knobs.h>.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>