[BuildingSystem] Add DMMLib support
This patch add the required entries to support the generation of an RTLib
with DMMLib integration.
If DMM support is enabled by KConfig options, beside the bbque_rtlib.so library
another shared oobject: bbque_rtlib_dmm.so is also generated.
This last version of the library is identical to the previous one but adds
DMM support, thus, all memory management calls (e.g. malloc, memalign and free)
will be handled at run-time by DMMLib.

Applications willing to exploit the DMMLib just need to link this second
version instead of the original one.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>