[Android] Add first JNI broker methods
This is a first, partial, implementation of the required code to bridge
the C/C++ native RTLib with a corresponding Java counterpart.
The RTLib representative at the Java side will be the BbqueService, a
class extending a basic Android Service with wrapper methods to interface
with the BarbequeRTRM run-time library.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>