[CommandDispatcher 1/4] Add initial implementation
This is the glue code required to provide a command dispatcher module.
This module is in charge to collect commands from different sources and
forward them to a proper executor.

A BBQ module which needs to expose a command is required to implement
the bbque::Commandhander interface, thus defining an implementation for
a CommandsCb method.

For the time being, just a FIFO channel is supported.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>