[PP] Add platform update notification
Once the platform description should be re-loaded, e.g. because of a
change on resources availability, this condition could be notified to
the PlatformProxy by calling the Refresh method.

The default implementation of this method is to simply notify the
worker task, which by default issues a new call to the
platform defined _RefreshPlatformData method.

Both PlatformProxy and Task implementations could be overwritten by
the platform specific proxy.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>