[PP_Linux] Add support for resources refresh at run-time
The first time the platform is parsed, all available resources are expected
to be defined by CGroups and registered to the ResourceAccounter.
After that time, the Linux platform proxy is switched into REFRESH MODE which
means that, if the availability of registered resources should change at
run-time, an UpdateResource is used to notify the ResourceAccounter about
the new availability.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>