[PP_Linux] Add support for Clusters OFF-LINING
Because of CGroups interface, we cannot assign an empty CPU quota.
The minimum allowed value is 1%, since this value is also quite
un-useful on a real configuration, we assume a CPU being offline
when a CPU quota <= 1% is required.

This has to be considered as a sort of "reasonable workaround" which is
specific of the Linux platform proxy.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>