[RB] Introduce class ResourceBinder
This is a static class providing methods to solve or provide information
about resource bindings.

The binding is a mandatory step, through which a scheduling policy links
the resources requested in the AWMs (referenced by a resource path built
from the recipe) to the system resources (referenced by the resource path
registered by the Platform Proxy.

Previously, class WorkingMode embedded code for performing this step.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>