[RP] Class ResourcePath
This introduces a class to build resource references. Previously a resource
path was intended as a string object. Query on the status of a resource
were performed by specifying this kind of string. By introducing an ad-hoc
class we can perform these queries in a more effective way, avoiding
operations on strings.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>