[RA] Use ResourcePath objects in registration
The PlatformProxy can still register resources specifying the path in the
char string format. Now the ResourceAccounter instantiates a ResourcePath
object for each registered resource, and put a shared pointer to it into
a map, where the key is the string. This also leads to a refactoring of
some data structures previously used to keep track of the number of
resources (per type) into the system.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>