[AWM] Resources data structures re-arrangement
Data structures used for managing resources information have been
re-structured.

Attributes of "resources" member data have been renamed as following, to
improve readability:
    from_recp -> requested
    on_sched  -> sched_bindings
    to_sync   -> sync_bindings

BindingInfo replaces ClustersInfo. A vector of this type of objects
("binding_masks") has been added to "resources" in order to keep track
of binding changed between the actual (synchronized) optimization, and
the previous. The vector is dimensioned on the number of resource types
defined in class ResourceIdentifier, since the resource binding can be
performed on whatever type of resource.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>