[Config] Binding domain parameter
The configuration parameter specify the resource binding domain, i.e. on
which resource path the scheduling policy should solve the translation of
resource IDs from the ones given in the recipe and the ones of the
registered resources.

Actually, this parameter defines a single binding domain. In the future
we could think about handling more than one domain.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>