[AM] Binding info on status table
Resource binding information are now printed out in a more readable
format. Anyway it is worthy to note that the binding domain here is
"hardcoded" on CPU resources. In the future, a more general schema
should be implemented, in order to be able to print out binding
information on whatever is the domain.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>