[P2012_PP] Fabric and PE quota refactoring
The fabric quota is now defined by the product: #clusters * #pes * 100.
Values related to number of clusters and pes are read from the attributes,
dynamically set by the platform runtime.

The PE fabric quota is the ratio between the quota specified into the
recipe and the fabric quota. This value, multiplied by 100 is then
written into the shared data structure, as the fabric quota constraint
assigned to the application.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>