[BuildSystem] Fix build for SThorm targets
When the target is the SThorm platform, due to the organization
of header files within the SThorm SDK, the flag __ZYNQ__ allows
to require a build for the physical platform (i.e. instead of the
GePOP simulator).

For the time being, we need just a portion of header files which
is independent of the distinction among simulated and real platform.
Thus, we force the usage of the physical one, which does not break
the compilation due to missing header file.

TODO: check with Germain for a better integration solution.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>