[P2012_PP] Workaround on message input interface
Apparently the P2012 device driver interface to get input messages is not
blocking, each call to p2012_getNextMessage returning immediately if no
new messages are available in the input queue.

This is dirty and bad workaround, polling is definitively not the solution!

TODO: add poll support to P2012 device driver.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>