[P2012_PP] Disabling un-supported code
For the time being, message based notifications are not supported
by the SThorm run-time. Using them just on the HOST side does
not allow to run BBQ for more than few schedule, since quite soon
the message queue is filled and a communication error is notified
by the device driver.

Thus, for the time being, a simple workaround is to disable
these messaging exchange.

TODO: better define the messaging interface

Signed-off-by: Patrick Bellasi <derkling@gmail.com>