[RTLIB_RR 2/3] Add SetThreshold API
The RTLib provides the support to profile and report the Reconfiguration
Rate (RR) of an application.

This patch add a new function which allows to define a threshold
value for the RR which should be satisfied in order to forward GoalGap
calls to the BarbequeRTRM.
At run-time, the RTLib monitors application behaviors, measure its
RR and silently block the forwarding of GoalGap calls in case the
RR is higher than the configured RR threshold value.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>