[P2012_PP] First power management support
This adds a first budget-based power management support.

A deferrable is instanced to sample the power consumption provided by the
platform. The sampling period can be modified through a command, as well as
the checking period, i.e., the time between a power management policy calls.

Such a policy, checks the current power consumption value (exponential
moving average) comparing it to a given (and run-time tunable) power
budget. Whenever the power budget is violated, the policy "reserves" a
certain amount of POWER resource and notify the Resource Manager a request
for a new optimization.

Conversely, if the current consumption is far below the current budget and
the total (unreserved) amount of resource POWER is below its registered max
(total), the policy slowly raise up again the unreserved amount, possibly
until resuming the original total amount.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>