[Tools_SignalAnalysis] Plot AWM stability
An AWM stability flag is added by this patch which is defined by
checking a condition on the Reconfiguration Rate.
This match with our definition of stability which allows an application
to reconfigure itself as long as the time spent on tuning is upper
bound limited.
The AWM stability flag is plotted with AWM metrics.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>