[P2012_PP] Validation of clusters/PEs count
This introduces a control on the number of clusters and processing elements
to validate the resource enumeration. If one of such values is not greater
than 0, it means that the STHORM run-time has not properly initialized the
data structure. Therefore the initialization, from the Barbeque side, is
interrupted and an error core is returned.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>