[Tools_SignalAnalysis] Refactor time tracking
This patch rafactor the time tracking system in order to track real
simulation time considering:
1. the actual time spent for onRun+onMonitor+onConfigure
2. the tracking of real simulation time

A simulation step is composed by predefined maximum number of
simulation cycles, than it is compensated to possibly be real-time.
Because of the computation overheads, right now the maximum achievable
real-time speed is about 6 CPS.
Most of the time is spent in the updateRR function, where the three
onXXX methods duration is estimated. This is the function to optimzie
in order to speed-up the real-time simulation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>