[RTLIB_RPC] Bugfix CPS enforcing initialization
Usually the SetCPS call is asserted in onConfigure calls, but according to
the current implementation the really first cycle after the onConfigure does
not benefits of the CPS enforcing since the cps_start time is initialized
in the PostMonitor notifier. Thus, the first cycle after the SetCPS could be
really short compared to the second one, resulting into an excessively high
variance in the SyncTime estimation code.

This patch properly initialize the cps_start time so that even the really
first cycle after an onConfigure will get the required cycle configured.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>