[RTLIB_RPC] Bugfix ForceCPS after reconfiguration
The ForceCPS code, which has been recently moved into the NotifyPreRun
in order to have a proper computation of the RR value, is now triggered
even in the really first onRun after a reconfiguration.
This introduces an unwanted pause in the first cycle after each
reconfiguration.

This patch solves that bug by exploiting the RR time_running value,
which is now reset after each (re)configuration thus disabling also
the ForceCPS call in the really first pre-onRun.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>