[RA 3/6] Keep sync lock on resources release
The synchronization look is required to serialize resources release with
synchronization activities in order to avoid misaligned representation on
resources availability when the sync view is reported on the system view.

NOTE: in in order to avoid dead-locks race, the sync view lock should
always be acquired before the status view.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>