[YM] Bugfix forced commit on PostChange
If an application is not to be sync, the PostChange should not enforce
the commit, otherwise, the following policy steps could be ignored being
the application remove from its synchronization queue.

TODO: a check on sync queue being empty should be eventually done after the
synchronization complete, before to switch the sync view to the system view.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>