[SC_Reconfig] Fix total amount resource query
The contribution needs to known at least one valid ID of a binding
domain (e.g., CPU node) to perform a query on a valid resource path, to
known the total amount of resource in the scope of a binding domain.
Such ID number is passed by the scheduling policy through the Init()
member function.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>