[SC_Fairness] Fix clustered fair partition computation
The contribution was incorrectly computed in the cases of unbalanced usage
of the computing nodes (e.g., CPUs).

For instance, if we have 2 CPUs with the following quota availability:
20, 20, 200, and 3 applications to schedule, the system-wide fair partition
is 240/3=80. However, the contribution was unable to allocate such quota,
since the "clustered" (binding domain) fair partition was seen equal to 20.
This patch fixes this mishbehavior.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>