[RTLIB] Allows AWM selection when running un UNMANAGED mode
The UNMANAGED mode flag "U" passed to the BBQUE_RTLIB_OPTS environment
variable could now be used to specify a custom AWM, different from the
default one which is 0.

Such flag could be followed by the ID of the AWM the application should be configured
to run into, for example:
  BBQUE_RTLIB_OPTS="U3"
will run the application in AWM 3.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>