[RB] Enable filtered binding
In the binding function has been enabled also the possibility to filter
out the set of resources that do not match a given filter criteria.
The filter is built according to the set of parameters of the SetResourcesList
member function of class Usage, i.e., the type of resource and the set of
IDs to match for inclusion.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>