[Tools_PluginDev] Plugin-development support
This introduces the first version of a script providing support for the
addition of new BarbequeRTRM plugins. NOTE: Current version allows the
addition of plugins of type 'schedpol' only.
The script instantiate new source files with a ready to compile template
of the plugin, and add it to the configuration and building system.

The tool can be launched by a BOSPShell by typing 'bbque-dev-plugin', and
providing the required arguments.
The first argument is the ACTION. Possible values are: create, delete, list,
help. In case of 'create' or 'delete' two further arguments must be
provided, the TYPE and the NAME of the plugin.
As previously said, the only TYPE actually manageable is 'schedpol'.
The NAME is the name of the new plugin (create), or of the plugin to remove
(delete).

Please type 'bbque-dev-plugin help' for a resume of the available actions.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>