[Config] Add a new configuration for max Resources ID number
Computation resources are assigned a unique ID number, which is counting
from 0 up to BBQUE_MAX_RID_NUMBER - 1. These IDs are used for example to
represent the CPUs of a target system or its memory nodes.
Usually this value is used to dimension bitmaps data structures.

This patch add a Kconfig defined value for this important parameter.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>