[RPC] Fire Forward Profiles even is app sleeps due to SetCPS enforcing
To do that, I created two cycle time accumulators instead of one.
The first is used by the application and contains the correct cycletime
values; the second is used by bbque and contains the would-be cycletime
values (i.e. do not take into account the sleeps). Bbque is thus aware
if cycletime is OK only because of the sleeps.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>

# Conflicts:
#	rtlib/bbque_rpc.cc