[SM] Don't wait for apps to reconfigure before reclaiming resources
After an application receives a new resource allocation, it
could choose to terminate before doing the reconfiguration.
When the application terminates, it notifies it to bbque.

This is a deadlock: bbque is waiting for the app to notify it
a configuration happened, while the application is waiting for
bbque to ACK the application termination.

Removing the PostChange response solves this problem.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>