[RT] Add the bbque option to retrieve the kernel configuration
Since there is no syscall to get the preemption model active in the
system, we have to search for the kernel configuration. Usually it
is in /proc/config.gz, but in some distributions this file is not
present due to performance reason. Ubuntu-like distributions place
/boot/config-* file with the full configuration.

This commit includes a bash parser of the kernel config and the
program_options call to retrieve the command line argument placed
to barbeque executable by the bash parser.

Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>