[RPC] Removed max CPS set on SetCPSGoal
Not all applications want their min cycletime to be enforced
when they declare a goal. Morevoer, runtime profile is now sent
when necessary (previously it was sent with a fixed frequency);
thus, resource overassignments are promptly corrected even without
cycletime enforcing.

The cycletime enforcing should be asked for implicitely,
or not used at all.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>