[RPC_DA] Trigger EXC reconfiguration on some allocation changes
In distributed actuation mode, the RTLib can dynamically tune the
EXC resource allocation, while staying in the BarbequeRTRM allocation
budget. EXC reconfiguration, however, was triggered only on budget change,
i.e., when a new AWM is assigned to the EXC by the BarbequeRTRM.

This commit introduced the possibility to trigger a reconfiguration even
if the AWM did not change. The reconfiguration is triggered if the
cpuset changes, or if the total exploitable pes number changes
(e.g. bandwidth changes from X to Y equivalent pes). In these cases,
in fact, the EXC may want to reconfigure, e.g. change the number of
active threads.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>