[RPC_DA] Adaptive cfs period for DA
In the case of RTLib CGroups Distributed Actuation, the length of
the cpu.cfs_period can be configured by menuconfig. For maximum
accuracy, this value should in fact be close to the real onRun duration.

The flag can be found in menuconfig, under DA options
(Res. management -> P. Setup -> Linux Res. Management -> CGroups)

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>