[RPC_DA] CPU resources release policy
When executing in DA mode, the RUntime Library can allocate to the
EXC less resources than those that have been allocated by the resoure
manager. Once the Runtime Library is sure that a given amount of resources
is definitely not needed by the EXC, these resources are released; that is,
the Runtime Library notifies to the resource manager that it can seize
these resources back.

The release policy can be selected in the menuconfig in the DA
configuration section. Policies define how many resources are released:

Aggressive: all the resources that are likely not to be needed
Conservative: part of the resources that are likely not to be used
Relaxed: nothing. If the resource manager will need some resources, it
   will seize them explicitly.


Signed-off-by: Simone Libutti <simone.libutti@gmail.com>