[RPC] Computing EXC cycletime using EMA
The Runtime Library computes the average cycletime of EXCs using
all the samples of the current working mode. That is, while the resource
allocation does not change, all the cycletime samples are used to
compute the average cycletime. This allows the Runtime Library to have
a more clear idea about how the current resource allocation is affecting
the EXC performance and about how the current allocation should be
modified to achieve an optimal performance.

Conversely, the EXC itself is mostly interested in its current
performance statistics. Given that the cycletime is subject to
variability, the value of the last cycle time (which can also be get
using the RTLib APIs) is not very informative; to tackle this issue,
the RTLib computes the Exponential Moving Average (EMA) of the last
N cycles.

The number of samples used to compute the EMA can be configured in
menuconfig, under:

   Runtime Resource Manager -> RTLib -> Performance API configuration

This commit introduces the above mentioned feature. The EMA computation
class was already present under the bbque::utils namespace. It was
previously used to compute the EXC performance (which, as said, is
computed through a 'standard' average in the current release).
This commit also removes the user cycletime analyser reset option from the
SetCPS method: it is not needed anymore, given that the EMA automatically
forgets its history.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>