[RPC_DA] Adapting number of cfs_periods for more accurate enforcement
The length of the cfs enforcement period has visible effects on the
enforcement accuracy: experimental results show that, depending on the
type of application (constant or variable CPU usage during the onRun),
different cfs enforcement periods can lead to resource under- or
over-usage.

This commit introduces an adaptive cfs period length heuristic that
selects the number of periods per cycle that maximize the enforcement
accuracy.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>