[RPC_DA] Committing cgroup allocation in case of bbque allocation change
Currently, while an EXC is satisfied with its performance, cgroups are
never updated.

However, if the global resource manager changes the resource allocation
for the EXC, it would make totally sense for the RTLib to apply the new
configuration as soon as possible (in DA, right before the next execution
cycle). This because other applications may have been allocated the
resources that the EXC is currently using.

This commit introduces this behavior.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>