[SchedPol_ManGA] Per partition AWM selection attempts
The AWM is selection/building process iterated over each possible
partition. The iteration stops after the first successful AWM scheduling
attempt.

The SelectBestPartition() function has been renamed into SortPartitions().
Current implementation does nothing but we expect to have a smarter
approach in the future.

The partition is set once the AWM has been actually assigned, as well as
the update of the overall mapping information in the task-graph.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>