[SM] Fixed erase() call invalidating iterator
In the for loop, the call to std::map::erase() was removing the object
pointed by iterator "resp_it" from the map structure, without updating
iterator's value.

At the end of the loop, the pre-increment was updating
the value of "resp_it", which was referring to the object previously
deleted.

Now the code retrieves the correctly updated value of the iterator
from the std::map::erase() call.